requirements.txt
nonebot2[fastapi]==2.3.2
nonebot-adapter-onebot==2.4.4
nonebot_plugin_alconna==0.49.0
pillow
opencv-python-headless
types-Pillow
requests
urllib3==1.26.5
SQLAlchemy
alembic
aiosqlite
pytz
qrcode
freezegun
imagetext-py
packaging